+++ EntraOps/Public/PrivilegedAccess/Save-EntraOpsPrivilegedEAMJson.ps1
@@ -24,8 +24,8 @@ function Save-EntraOpsPrivilegedEAMJson {
         [System.String]$ExportFolder = $DefaultFolderClassifiedEam
         ,
         [Parameter(Mandatory = $False)]
-        [ValidateSet("EntraID", "IdentityGovernance", "DeviceManagement", "ResourceApps")]
-        [Array]$RbacSystems = ("EntraID", "IdentityGovernance", "DeviceManagement", "ResourceApps")
+        [ValidateSet("EntraID", "IdentityGovernance", "DeviceManagement", "ResourceApps", "Defender")]
+        [Array]$RbacSystems = ("EntraID", "IdentityGovernance", "DeviceManagement", "ResourceApps", "Defender")
     )
 
     Write-Output "Clearing cache before analyzing RBAC and classification data"
@@ -38,8 +38,7 @@ function Save-EntraOpsPrivilegedEAMJson {
         if ((Test-Path -path "$($ExportFolder)")) {
             Remove-Item "$($ExportFolder)" -Force -Recurse
             New-Item "$($ExportFolder)" -ItemType Directory -Force
-        }
-        else {
+        } else {
             New-Item "$($ExportFolder)" -ItemType Directory -Force
         }
 
@@ -64,8 +63,7 @@ function Save-EntraOpsPrivilegedEAMJson {
         if ((Test-Path -path "$($ExportFolder)")) {
             Remove-Item "$($ExportFolder)" -Force -Recurse
             New-Item "$($ExportFolder)" -ItemType Directory -Force
-        }
-        else {
+        } else {
             New-Item "$($ExportFolder)" -ItemType Directory -Force
         }
 
@@ -92,8 +90,7 @@ function Save-EntraOpsPrivilegedEAMJson {
         if ((Test-Path -path "$($ExportFolder)")) {
             Remove-Item "$($ExportFolder)" -Force -Recurse
             New-Item "$($ExportFolder)" -ItemType Directory -Force
-        }
-        else {
+        } else {
             New-Item "$($ExportFolder)" -ItemType Directory -Force
         }
         $EamDeviceMgmt | Convertto-Json -Depth 10 | Out-File -Path "$($ExportFolder)/DeviceManagement.json"
@@ -119,12 +116,36 @@ function Save-EntraOpsPrivilegedEAMJson {
         if ((Test-Path -path "$($ExportFolder)")) {
             Remove-Item "$($ExportFolder)" -Force -Recurse
             New-Item "$($ExportFolder)" -ItemType Directory -Force
-        }
-        else {
+        } else {
             New-Item "$($ExportFolder)" -ItemType Directory -Force
         }
         $EamIdGov | Convertto-Json -Depth 10 | Out-File -Path "$($ExportFolder)/IdentityGovernance.json"
         foreach ($PrivilegedObject in $EamIdGov) {
+            $ObjectType = $PrivilegedObject.ObjectType
+            $SingleJSONExportPath = "$($ExportFolder)/$ObjectType"
+            If (!(test-path $SingleJSONExportPath)) {
+                New-Item -ItemType Directory -Force -Path $SingleJSONExportPath
+            }
+            $PrivilegedObject | Convertto-Json -Depth 10 | Out-File -Path "$SingleJSONExportPath/$($PrivilegedObject.ObjectId).json" -Force
+        }
+    }
+    #endregion
+    #region Defender
+    if ($RbacSystems -contains "Defender") {
+        $ExportFolder = "$($DefaultFolderClassifiedEam)/Defender"
+
+        $EamDefender = Get-EntraOpsPrivilegedEAMDefender
+        $EamDefender | Measure-Object
+        $EamDefender = $EamDefender | where-object { $null -ne $_.ObjectType -and $null -ne $_.ObjectId }
+
+        if ((Test-Path -path "$($ExportFolder)")) {
+            Remove-Item "$($ExportFolder)" -Force -Recurse
+            New-Item "$($ExportFolder)" -ItemType Directory -Force
+        } else {
+            New-Item "$($ExportFolder)" -ItemType Directory -Force
+        }
+        $EamDefender | Convertto-Json -Depth 10 | Out-File -Path "$($ExportFolder)/Defender.json"
+        foreach ($PrivilegedObject in $EamDefender) {
             $ObjectType = $PrivilegedObject.ObjectType
             $SingleJSONExportPath = "$($ExportFolder)/$ObjectType"
             If (!(test-path $SingleJSONExportPath)) {