+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 # Change Log
 All essential changes on EntraOps will be documented in this changelog.
 
+## [0.4] - 2025-05-30
+### Added
+- Support for Role Management Provider "Defender" (Unified RBAC for Microsoft Defender XDR)
+  - Currently, the API does not include details on Device Groups or Scope. Therefore, the RBAC system is not covered by using default settings (EntraOps.config) to avoid wrong classification by missing consideration of scope.
+
 ## [0.3.4] - 2024-12-21
 ### Fixed
 - Type of Owners field is inconsistent [#31](https://github.com/Cloud-Architekt/EntraOps/issues/31)